revert unhandledRejection hook fix from RD-9130
